# BiuroPodrozy.Data/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BiuroPodrozyAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Cities": {
      "get": {
        "tags": [
          "Cities"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CityDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cities"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Cities/{id}": {
      "get": {
        "tags": [
          "Cities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CityDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cities"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Countries": {
      "get": {
        "tags": [
          "Countries"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Countries"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Countries/{id}": {
      "get": {
        "tags": [
          "Countries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Countries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CountryDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Countries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CityDTO": {
        "type": "object",
        "properties": {
          "idCity": {
            "type": "integer",
            "format": "int32"
          },
          "cityName": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idCountry": {
            "type": "integer",
            "format": "int32"
          },
          "countryName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CountryDTO": {
        "type": "object",
        "properties": {
          "idCountry": {
            "type": "integer",
            "format": "int32"
          },
          "countryName": {
            "type": "string",
            "nullable": true
          },
          "abbrCountryName": {
            "type": "string",
            "nullable": true
          },
          "continent": {
            "type": "string",
            "nullable": true
          },
          "population": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "countrySize": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "phoneCode": {
            "type": "string",
            "nullable": true
          },
          "idCurrency": {
            "type": "integer",
            "format": "int32"
          },
          "currencyName": {
            "type": "string",
            "nullable": true
          },
          "idLanguage": {
            "type": "integer",
            "format": "int32"
          },
          "languageName": {
            "type": "string",
            "nullable": true
          },
          "idCapitalCity": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "capitalCityName": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}